requirements.txt
Django==1.7.4
django-cors-headers==1.0.0
djangorestframework==3.0.4
wsgiref==0.1.2
